MaatwerkcodeAzureFunctions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
boto3==1.14.20
azure-storage-blob
cjvalpy[validate]
lxml